+++ react/SearchResult.js
@@ -56,27 +56,6 @@ export default class SearchResult extends Component {
     }
   }
 
-  get breadcrumbsProps() {
-    const {
-      params: { category, department, term },
-    } = this.props
-
-    const categories = []
-
-    if (department) {
-      categories.push(department)
-    }
-
-    if (category) {
-      categories.push(`${department}/${category}/`)
-    }
-
-    return {
-      term,
-      categories,
-    }
-  }
-
   render() {
     return (
       <SearchResultInfiniteScroll
